refactoring the perception wrapper code
